requirements.txt
pycryptodome
requests
reportlab
termcolor